test(maestro): one in-memory LedgerStore instead of twelve
memoryStore() was pasted into 12 test modules and had drifted into 9 "distinct"
bodies. Almost all of the divergence was cosmetic — `record` vs `r`, hoisted vs
inline filter locals — which is exactly the kind of noise that hides the four
differences that were real: direct access to the backing collections (under
three different accessor names: sessions/queries, dump/rows, raw), the optional
putSessionIf compare-and-set seam, microtask-hop injection for interleaving
tests, and a one-shot putSession failure.

tests/helpers/memory-store.ts folds all four in as options/extras. With no
options the behaviour is the plain copies' exactly. Dead type imports left by
the removals are dropped too.

421 maestro tests green, testbed 37 green, dependency-direction guard OK.
