+++ projects/silver-core-maestro-sdk/tests/ledger.test.ts
@@ -1,40 +1,8 @@
 import { describe, it, expect } from 'vitest';
 import { TaskLedger, DuplicateSessionError } from '../src/ledger/ledger.js';
 import { InvalidTransitionError } from '../src/ledger/state.js';
-import type { LedgerStore, SessionFilter } from '../src/ledger/store.js';
-import type { QueryRecord, SessionRecord } from '../src/ledger/types.js';
-
-/** Minimal in-memory store for unit tests (hosts inject their own). */
-function memoryStore(): LedgerStore & { sessions: Map<string, SessionRecord> } {
-  const sessions = new Map<string, SessionRecord>();
-  const queries: QueryRecord[] = [];
-  return {
-    sessions,
-    async putSession(record) {
-      sessions.set(record.id, { ...record });
-    },
-    async getSession(id) {
-      const r = sessions.get(id);
-      return r === undefined ? null : { ...r };
-    },
-    async listSessions(filter?: SessionFilter) {
-      let all = [...sessions.values()];
-      if (filter?.states !== undefined) {
-        all = all.filter((s) => filter.states!.includes(s.state));
-      }
-      if (filter?.dueBefore !== undefined) {
-        all = all.filter((s) => s.nextRunAt !== null && s.nextRunAt <= filter.dueBefore!);
-      }
-      return all.map((s) => ({ ...s }));
-    },
-    async appendQuery(record) {
-      queries.push({ ...record });
-    },
-    async listQueries(sessionId) {
-      return queries.filter((q) => q.sessionId === sessionId).map((q) => ({ ...q }));
-    },
-  };
-}
+import type { LedgerStore } from '../src/ledger/store.js';
+import { memoryStore } from './helpers/memory-store.js';
 
 function testClock(startAt = 1_000) {
   let t = startAt;
