+++ projects/silver-core-maestro-sdk/tests/property-ledger.test.ts
@@ -10,38 +10,7 @@ import { describe, it, expect } from 'vitest';
 import fc from 'fast-check';
 import { TaskLedger, DuplicateSessionError } from '../src/ledger/ledger.js';
 import { InvalidTransitionError } from '../src/ledger/state.js';
-import type { LedgerStore, SessionFilter } from '../src/ledger/store.js';
-import type { QueryRecord, SessionRecord } from '../src/ledger/types.js';
-
-function memoryStore(): LedgerStore & { dump: () => SessionRecord[]; rows: () => QueryRecord[] } {
-  const sessions = new Map<string, SessionRecord>();
-  const queries: QueryRecord[] = [];
-  return {
-    dump: () => [...sessions.values()].map((s) => ({ ...s })),
-    rows: () => queries.map((q) => ({ ...q })),
-    async putSession(r) {
-      sessions.set(r.id, { ...r });
-    },
-    async getSession(id) {
-      const r = sessions.get(id);
-      return r === undefined ? null : { ...r };
-    },
-    async listSessions(filter?: SessionFilter) {
-      let all = [...sessions.values()];
-      if (filter?.states !== undefined) all = all.filter((s) => filter.states!.includes(s.state));
-      if (filter?.dueBefore !== undefined) {
-        all = all.filter((s) => s.nextRunAt !== null && s.nextRunAt <= filter.dueBefore!);
-      }
-      return all.map((s) => ({ ...s }));
-    },
-    async appendQuery(r) {
-      queries.push({ ...r });
-    },
-    async listQueries(sessionId) {
-      return queries.filter((q) => q.sessionId === sessionId).map((q) => ({ ...q }));
-    },
-  };
-}
+import { memoryStore } from './helpers/memory-store.js';
 
 type Op =
   | { op: 'dispatch'; key: number; manual: boolean; maxAttempts: number; delay: number }
